# funcr: Add LogInfoLevel Option to skip logging level in the info log