--- a/funcr/funcr_test.go
+++ b/funcr/funcr_test.go
@@ -1361,3 +1361,35 @@ func TestOptionsTimestampFormat(t *testing.T) {
 		t.Errorf("\nexpected %q\n     got %q", expect, capt.log)
 	}
 }
+
+func TestOptionsLogInfoLevel(t *testing.T) {
+	testCases := []struct {
+		name   string
+		level  string
+		expect string
+	}{
+		{
+			name:   "custom key",
+			level:  "info_level",
+			expect: `"info_level"=0 "msg"="msg"`,
+		},
+		{
+			name:   "no level",
+			level:  "",
+			expect: `"msg"="msg"`,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run("Run: "+tc.name, func(t *testing.T) {
+			capt := &capture{}
+			sink := newSink(capt.Func, NewFormatter(Options{LogInfoLevel: &tc.level}))
+			dSink, _ := sink.(logr.CallDepthLogSink)
+			sink = dSink.WithCallDepth(1)
+			sink.Info(0, "msg")
+			if capt.log != tc.expect {
+				t.Errorf("\nexpected %q\n     got %q", tc.expect, capt.log)
+			}
+		})
+	}
+}